// src/app/component/model/Reservation.ts
export interface Reservation {
  id?: number;                // Optional because it is generated by the backend
  clientId: string;           // Client ID who made the reservation
  apartmentId: number;       // ID of the apartment being reserved
  startDate: string;         // Start date of the reservation in 'YYYY-MM-DD' format
  endDate: string;           // End date of the reservation in 'YYYY-MM-DD' format
  totalPrice: number;
  status: 'PENDING' | 'ACCEPTED' | 'REJECTED'; // Reservation status
  createdAt: string;         // The creation time of the reservation in 'YYYY-MM-DDTHH:mm:ss' format
}